Add modifications.
Added a split feature that allows the user to split the total evenly among 1-15 people.
Added a round up feature that allows the user to round up the total to the nearest
dollar. This will update the tip and tip % automatically while keeping the base amount.
When the toggle is switched back to off, the old tip % will be remembered and restored.
Began working on a currency feature. Currently have a Spinner menu in place that can
switch between 6 different currencies, but it has no effect on the numbers.