Implement history feature.
Added support for the user to save payments within a current session and view their
history. Clicking the View History button transitions to a new HistoryActivity, which
contains a RecyclerView widget to display all of the transactions in a scrolling view.
However, this implementation is limited in that the data is not persistent over time.